Completely reworked meager.io
This is for the better - now that I'm actually capable of working on side
projects (yay anti-depressants) this is a project-focused website. I
know that I won't make blazing-fast progress, but there's an opportunity
to keep things just a little more updated and pick at multiple small
things. Should be fun. Also 80 columns GET.